+++ src/shared/component/header/Tabs.tsx
@@ -1,0 +1,138 @@
+// 로그인 여부에 따라 보여지는 탭 변경
+import Image from "next/image";
+import Link from "next/link";
+import { useRouter } from "next/router";
+import React from "react";
+
+interface TabsProps {
+  isLoggedIn: boolean;
+  headerColor: string;
+}
+
+const Tabs = ({ isLoggedIn, headerColor }: TabsProps) => {
+  const router = useRouter();
+
+  return (
+    <div className="gap-12 flex-center">
+      {isLoggedIn ? (
+        // 유저가 로그인한 경우
+        <>
+          {loginedtabs.map((tab, i) => (
+            <Link key={`tab-${i}`} href={tab.path}>
+              <div
+                className={`text-lg font-medium ${
+                  headerColor === "white" ? "text-gray-500" : "text-white"
+                } cursor-pointer hover:underline`}
+              >
+                {tab.title}
+              </div>
+            </Link>
+          ))}
+          <div className="flex items-center">
+            <Image
+              className="cursor-pointer"
+              src="/header/notification.svg"
+              alt="notification"
+              width={25}
+              height={28}
+            />
+          </div>
+          <div
+            className={`flex ${
+              isLoggedIn ? "w-[198px]" : "w-[303px]"
+            }  h-[62px] bg-white rounded-[34.50px] shadow p-1 items-center`}
+          >
+            <div className="flex items-center gap-3 px-4">
+              <div className="w-[94px] text-center text-gray-500 text-xl font-bold font-['Roboto'] capitalize truncate ...">
+                Nickname
+              </div>
+              <div className="flex-center">
+                <Image
+                  src="/header/photo_sample.svg"
+                  alt="profile img"
+                  width={54}
+                  height={54}
+                />
+              </div>
+            </div>
+          </div>
+        </>
+      ) : (
+        // 유저가 로그인하지 않은 경우
+        <>
+          {defaultTabs.map((tab, i) => (
+            <Link key={`tab-${i}`} href={tab.path}>
+              <div
+                className={`text-lg font-medium ${
+                  headerColor === "white" ? "text-gray-500" : "text-white"
+                } cursor-pointer hover:underline`}
+              >
+                {tab.title}
+              </div>
+            </Link>
+          ))}
+          <div
+            className={`flex ${
+              isLoggedIn ? "w-[198px]" : "w-[303px]"
+            } h-[62px] bg-white rounded-[34.50px] shadow p-1 items-center`}
+          >
+            {loginTabs.map((loginTab, i) => (
+              <Link key={`loginTab-${i}`} href={loginTab.path}>
+                <div
+                  className={`w-[145px] h-[49px] ${
+                    router.asPath === loginTab.path
+                      ? "bg-orange-100 hover:bg-orange-200"
+                      : "bg-white hover:bg-gray-200"
+                  } rounded-[34.50px] p-1 ml-2 cursor-pointer flex-center`}
+                >
+                  <span className="w-[70px] h-[25px] text-center text-gray-500 text-xl font-bold font-['Roboto'] capitalize">
+                    {loginTab.title}
+                  </span>
+                </div>
+              </Link>
+            ))}
+          </div>
+        </>
+      )}
+    </div>
+  );
+};
+
+export default Tabs;
+
+const loginedtabs = [
+  {
+    title: "일정 찾기",
+    path: "/schedule",
+  },
+  {
+    title: "일정 만들기",
+    path: "/create-schedule",
+  },
+  {
+    title: "내 일정",
+    path: "/my-schedule",
+  },
+];
+
+const defaultTabs = [
+  {
+    title: "일정 찾기",
+    path: "/schedule",
+  },
+  {
+    title: "일정 만들기",
+    path: "/create-schedule",
+  },
+];
+
+const loginTabs = [
+  {
+    title: "Log in",
+    path: "/login",
+  },
+  {
+    title: "sign up",
+    path: "/signup",
+  },
+];
